pyproject.toml
[tool.poetry]
name = "ramadanplanner-backend"
version = "0.1.0"
description = ""
authors = ["Abdul Ahad <abdul.ahad@brainstation-23.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.2"
django-ninja-extra = "^0.20.4"
python-dotenv = "^1.0.1"
django-cors-headers = "^4.3.1"
psycopg2-binary = "^2.9.9"
pyjwt = "^2.8.0"
pydantic = {extras = ["email"], version = "^2.6.1"}
hijridate = "^2.4.1"
pillow = "^10.2.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.1"
mypy = "^1.8.0"
django-stubs = "^4.2.7"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
